requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery-datatransfer>=3.0.0
requests>=2.31.0
urllib3>=2.0.0
